styling updated on the sign in and create acc page. Homepage updated. CreateAcc component created.
