+++ client/src/components/shared/Nav/Nav.jsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import './Nav.css'
 import { Link } from 'react-router-dom'
-import { Navbar, Nav} from 'react-bootstrap'
+import { Navbar, Nav } from 'react-bootstrap'
+
 
 export default function Navigation() {
 
@@ -10,7 +11,7 @@ export default function Navigation() {
     <Navbar collapseOnSelect expand="sm" className='custom-nav' variant='light' sticky='top'>
       <Navbar.Brand><Link className='text-dark custom-font' to='/'>Breathe</Link></Navbar.Brand>
       <Navbar.Toggle aria-controls="responsive-navbar-nav" />
-      
+
       <Navbar.Collapse id="responsive-navbar-nav">
         <Nav className="m-auto">
           <Link className='nav-link text-dark' to='/'>My Account</Link>
@@ -19,7 +20,7 @@ export default function Navigation() {
           <Link className='nav-link text-dark' to='/'>Explore Services</Link>
           <Link className='nav-link text-dark' to='/'>Confidentiality Policy</Link>
         </Nav>
-      </Navbar.Collapse> 
+      </Navbar.Collapse>
     </Navbar>
   )
 }
