+++ client/src/components/shared/SignIn/Signin.css
@@ -1,13 +1,16 @@
 .sign-in-ctn{
   display: flex;
   justify-content: center;
-  border-right: 3px solid black;
+  /* border-right: 3px solid black; */
   width: 50%;
-  
+  margin-top: 15px;
+  margin-bottom: 50px;
 }
 
 .inner-ctn{
- width: 50%;
+ width: 75%;
+ border-right: 3px solid black;
+ margin-left: 50px;
 }
 .user-sign{
 width: 900px;
@@ -21,9 +24,10 @@ height: 200px;
   border: 1px solid rgb(16, 202, 143);
 }
 .sign-in{
-  font-size: 18px;
-  width: 150px;
-  margin: 0 auto;
+  font-size: 24px;
+  width: 250px;
+  margin: 0 auto 0 90px;
+  padding-bottom: 35px;
 }
 .sign-btn{
   width: 280px;
@@ -33,5 +37,5 @@ height: 200px;
   border-radius: 25px;
   border : 1px solid rgb( 177, 75, 190);
   background-color: rgb( 177, 75, 190); 
-  margin: -50px 0 0 12px;
+  margin: 5px 0 0 12px;
 }
