+++ client/src/screens/UserCreate/UserCreate.css
@@ -1,17 +1,13 @@
 .nav-img{
   height: 150px;
-  width: 100%;
   
 }
 /* .create-acc-ctn{
   margin-left: 65%;
   width: 50%;
-  
 }
 .form-ctn{
- 
 } */
-
 .outer-ctn{
   display: flex;
   flex-flow: row-reverse;
